+++ eslint.config.mjs
@@ -16,6 +16,10 @@ const eslintConfig = defineConfig([
     "build/**",
     "next-env.d.ts",
   ]),
-]);
+], {
+  rules: {
+    "@typescript-eslint/no-explicit-any": "off",
+  },
+});
 
 export default eslintConfig;